Restructure project | Polish templates | Add TodoTask test

+++ app/todo/views.py
@@ -1,51 +1,29 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
-from django.contrib.auth import get_user_model, authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import ObjectDoesNotExist
 
 from .models import TodoTask
 from .forms import TodoTaskForm
-
-
-def loginPage(request):
-
-    if request.method == "POST":
-        email = request.POST.get("email")
-        password = request.POST.get("password")
-
-        try:
-            user = get_user_model().objects.get(email=email)
-
-            user = authenticate(request, email=email, password=password)
-
-            if user is not None:
-                login(request, user)
-                return redirect("home")
-            else:
-                messages.error(request, "Invalid credentials")
-        except:
-            messages.error(request, "Invalid credentials")
-
-    context = {}
-    return render(request, "todo/login_register.html", context)
-
-
-def logoutUser(request):
-    logout(request)
-    return redirect("home")
+from core.views import errorPage
 
 
 def home(request):
-    q = request.GET.get("q") if request.GET.get("q") != None else ""
-    todo_tasks = TodoTask.objects.filter(
-        Q(name__icontains=q) | Q(description__icontains=q)
-    )
-    todo_tasks_count = todo_tasks.count
+    context = {}
 
-    context = {
-        "todo_tasks": todo_tasks,
-        "todo_tasks_count": todo_tasks_count,
-    }
+    if request.user.is_authenticated:
+        q = request.GET.get("q") if request.GET.get("q") != None else ""
+        todo_tasks = TodoTask.objects.filter(
+            Q(name__icontains=q) | Q(description__icontains=q)
+        ).filter(owner=request.user)
+        todo_tasks_count = todo_tasks.count
+
+        context = {
+            "todo_tasks": todo_tasks,
+            "todo_tasks_count": todo_tasks_count,
+        }
+
     return render(request, "todo/home.html", context)
 
 
@@ -55,22 +33,39 @@ def todoTask(request, pk):
     return render(request, "todo/todo_task.html", context)
 
 
+@login_required(login_url="login")
 def createTodoTask(request):
     form = TodoTaskForm()
 
     if request.method == "POST":
         form = TodoTaskForm(request.POST)
         if form.is_valid():
-            form.save()
+            todoTask = form.save(commit=False)
+            todoTask.owner = request.user
+            todoTask.save()
             return redirect("home")
 
     context = {"form": form}
     return render(request, "todo/todo_task_form.html", context)
 
 
+@login_required(login_url="login")
 def updateTodoTask(request, pk):
-    todo_task = TodoTask.objects.get(id=pk)
+    try:
+        todo_task = TodoTask.objects.get(id=pk)
+    except ObjectDoesNotExist:
+        return errorPage(
+            request,
+            err_msg="Task Not Found",
+        )
+
     form = TodoTaskForm(instance=todo_task)
+
+    if request.user != todo_task.owner:
+        return errorPage(
+            request,
+            err_msg="You are not authorized to edit this task",
+        )
 
     if request.method == "POST":
         form = TodoTaskForm(request.POST, instance=todo_task)
@@ -82,8 +77,21 @@ def updateTodoTask(request, pk):
     return render(request, "todo/todo_task_form.html", context)
 
 
+@login_required(login_url="login")
 def deleteTodoTask(request, pk):
-    todo_task = TodoTask.objects.get(id=pk)
+    try:
+        todo_task = TodoTask.objects.get(id=pk)
+    except ObjectDoesNotExist:
+        return errorPage(
+            request,
+            err_msg="Task Not Found",
+        )
+
+    if request.user != todo_task.owner:
+        return errorPage(
+            request,
+            err_msg="You are not authorized to delete this task",
+        )
 
     if request.method == "POST":
         todo_task.delete()
